Add files via upload
This notebook segmentate the texts in the documents from 'TextExtraction_Git' notebook excel file. The segmentation is perform using the tags from the pages in every document, this allow us to save metadata and retrieve the reference from the text. The tag used for segmentation can be changed or even we can extract all the text and perform random text segmentation.